requirements.txt
Cython==0.29.21
h5py==2.10.0
imgaug==0.4.0
ipython==7.16.1
ipython-genutils==0.2.0
Keras==2.0.8
matplotlib==3.3.3
numpy==1.19.2
opencv-python==4.5.1.48
Pillow==8.1.0
scikit-image==0.17.2
scipy==1.5.2
tensorflow==1.14.0